Imported Upstream version 0.9.5

+++ src/OverlayUnidraw/ovimport.c
@@ -114,6 +114,10 @@ implementPtrList(FileList,FILE)
 implementPtrList(StreamList,istream)
 implementPtrList(HandlerList,ReadImageHandler)
 
+#if __GNUG__>=3
+static char newline;
+#endif
+
 
 /*****************************************************************************/
 
@@ -465,6 +469,7 @@ int ReadImageHandler::process(const char* newdat, int len) {
 #else
       char buffer[BUFSIZ];
       in.get(buffer, BUFSIZ);
+      in.get(newline);
 #endif
 
       if (strncmp(buffer, "P6", 2)) {
@@ -478,6 +483,7 @@ int ReadImageHandler::process(const char* newdat, int len) {
         in.gets(&buffer);
 #else
 	in.get(buffer,BUFSIZ);
+	in.get(newline);
 #endif
       } while (buffer[0] == '#');
       sscanf(buffer, "%d %d", &width, &height);
@@ -486,6 +492,7 @@ int ReadImageHandler::process(const char* newdat, int len) {
       in.gets(&buffer);
 #else
       in.get(buffer,BUFSIZ);
+      in.get(newline);
 #endif
       int maxval;
       sscanf(buffer, "%d", &maxval);
@@ -609,9 +616,6 @@ int ReadImageHandler::inputReady(int fd) {
     GraphicComp* comp = OvImportCmd::DoImport(
       *ifs, empty, _helper, _ed, true, _path, newfd
     );
-#if __GNUG__>=3
-    if (ifptr) fclose(ifptr);
-#endif
 
 #if defined(OPEN_DRAWTOOL_URL)
     if (comp && comp->IsA(OVERLAY_IDRAW_COMP)) {
@@ -756,7 +760,7 @@ FILE* OvImportCmd::CheckCompression(
 
 
 const char* OvImportCmd::ReadCreator (const char* pathname) {
-    FILE* file = fopen(pathname, "r");
+    FILE* file = fopen(pathname, "r+");
     const int creator_size = 32;
     static char creator[creator_size];
 
@@ -803,6 +807,9 @@ const char* OvImportCmd::ReadCreator (const char* pathname) {
 	    else if (CheckMagicBytes(JPEG_MAGIC_BYTES, line)) 
 	      strncpy(creator, "JPEG", creator_size);
     
+	    else if (strncmp(line, "\211PNG", 4)==0)
+	      strncpy(creator, "PNG", creator_size);
+
             /* One-byte Magic numbers */
 	    else {
 		switch (line[0]) {
@@ -928,6 +935,9 @@ const char* OvImportCmd::ReadCreator (istream& in, FileType& ftype) {
     else if (CheckMagicBytes(JPEG_MAGIC_BYTES, line)) 
       strncpy(creator, "JPEG", creator_size);
     
+    else if (strncmp(line, "\211PNG", 4)==0)
+      strncpy(creator, "PNG", creator_size);
+
     /* One-byte Magic numbers */
     else {
       switch (line[0]) {
@@ -1444,6 +1454,8 @@ GraphicComp* OvImportCmd::Import (istream& instrm, boolean& empty) {
     GraphicComp* comp = nil;
     pnmfd = -1;
     OverlayCatalog* catalog = (OverlayCatalog*)unidraw->GetCatalog();
+    static boolean dithermap_flag = catalog->GetAttribute("dithermap") &&
+      strcmp(catalog->GetAttribute("dithermap"),"false")!=0;
 
     int len = 255;
     char buf[len];
@@ -1518,9 +1530,9 @@ GraphicComp* OvImportCmd::Import (istream& instrm, boolean& empty) {
 	  if (pathname && !return_fd) {
 	    char buffer[BUFSIZ];
 	    if (compressed) 
-	      sprintf(buffer, "gzip -c %s | pstoedit -f idraw", pathname);
+	      sprintf(buffer, "gzip -c %s | pstoedit -f idraw", pathname, "%d");
 	    else
-	      sprintf(buffer, "pstoedit -f idraw %s", pathname);
+	      sprintf(buffer, "pstoedit -f idraw %s", pathname, "%d");
 	    pptr = popen(buffer, "r");
 	    cerr << "input opened with " << buffer << "\n";
 	    if (pptr) 
@@ -1532,14 +1544,12 @@ GraphicComp* OvImportCmd::Import (istream& instrm, boolean& empty) {
           new_in.rdbuf()->attach(new_fd);
 #else
 	  FILE* ifptr = fdopen(new_fd, "r");
+	  helper.add_file(ifptr);
 	  filebuf fbuf(ifptr, ios_base::in);
 	  istream new_in(&fbuf);
 #endif
 	  comp = catalog->ReadPostScript(new_in);
 	  if (pptr) pclose(pptr);
-#if __GNUG__>=3
-	  if (ifptr) fclose(ifptr);
-#endif
 	} else
 	  cerr << "pstoedit not found\n";
       }
@@ -1608,17 +1618,17 @@ GraphicComp* OvImportCmd::Import (istream& instrm, boolean& empty) {
 
 	if (pathname && !return_fd) {
 	  char buffer[BUFSIZ];
-#if 0
-	  if (compressed) 
-	    sprintf(buffer, "cm=`ivtmpnam`;stdcmapppm>$cm;gzip -c %s | djpeg -map $cm -dither fs -pnm;rm $cm", pathname);
-	  else
-	    sprintf(buffer, "cm=`ivtmpnam`;stdcmapppm>$cm;djpeg -map $cm -dither fs -pnm %s;rm $cm", pathname);
-#else
-	  if (compressed) 
-	    sprintf(buffer, "cm=`ivtmpnam`;stdcmapppm>$cm;gzip -c %s | djpeg  -pnm;rm $cm", pathname);
-	  else
-	    sprintf(buffer, "cm=`ivtmpnam`;stdcmapppm>$cm;djpeg -pnm %s;rm $cm", pathname);
-#endif
+	  if (dithermap_flag) {
+	    if (compressed) 
+	      sprintf(buffer, "cm=`ivtmpnam`;stdcmapppm>$cm;gzip -c %s | djpeg -map $cm -dither fs -pnm;rm $cm", pathname);
+	    else
+	      sprintf(buffer, "cm=`ivtmpnam`;stdcmapppm>$cm;djpeg -map $cm -dither fs -pnm %s;rm $cm", pathname);
+	  } else {
+	    if (compressed) 
+	      sprintf(buffer, "gzip -c %s | djpeg  -pnm", pathname);
+	    else
+	      sprintf(buffer, "djpeg -pnm %s", pathname);
+	  }
 	  FILE* pptr = popen(buffer, "r");
           helper.add_pipe(pptr);
 	  if (pptr) {
@@ -1633,12 +1643,42 @@ GraphicComp* OvImportCmd::Import (istream& instrm, boolean& empty) {
             helper.add_stream(new_in);
 	    comp = PNM_Image(*new_in);
 	  }
-	} else
-	  comp = PNM_Image_Filter(*in, return_fd, pnmfd, "cm=`ivtmpnam`;stdcmapppm>$cm;djpeg -map $cm -dither fs -pnm;rm $cm");
+	} else {
+	  if (dithermap_flag) 
+	    comp = PNM_Image_Filter(*in, return_fd, pnmfd, "cm=`ivtmpnam`;stdcmapppm>$cm;djpeg -map $cm -dither fs -pnm;rm $cm");
+	  else 
+	    comp = PNM_Image_Filter(*in, return_fd, pnmfd, "djpeg -pnm");
+	}
       } else
 	cerr << "djpeg or stdcmapppm not found\n";
 
+    } else if (strncmp(creator, "PNG", 3)==0) {
+      if (OverlayKit::bincheck("pngtopnm")) {
+	if (pathname && !return_fd) {
+	  char buffer[BUFSIZ];
+	  if (compressed)
+	    sprintf(buffer, "gzip -c %s | pngtopnm", pathname);
+	  else
+	    sprintf(buffer, "pngtopnm %s", pathname);
+	  FILE* pptr = popen(buffer, "r");
+	  if (pptr) {
+	    cerr << "input opened with " << buffer << "\n";
+#if __GNUG__<3
+	    ifstream new_in;
+            new_in.rdbuf()->attach(fileno(pptr));
+#else
+	    filebuf fbuf(pptr, ios_base::in);
+	    istream new_in(&fbuf);
+#endif
+	    comp = PNM_Image(new_in);
+	    pclose(pptr);
+	  }
+	} else	
+	  comp = PNM_Image_Filter(*in, return_fd, pnmfd, "pngtopnm");
+      } else 
+	cerr << "pngtopnm not found\n";
     }
+
 
     if (comp && comp->IsA(OVRASTER_COMP)) 
       ((RasterOvComp*)comp)->SetByPathnameFlag(false);
@@ -1892,6 +1932,7 @@ OverlayRaster* OvImportCmd::PGM_Raster (istream& in, boolean ascii) {
 #else
     char buffer[BUFSIZ];
     in.get(buffer, BUFSIZ);
+    in.get(newline);
 #endif
 
     do {  // CREATOR and other comments
@@ -1899,6 +1940,7 @@ OverlayRaster* OvImportCmd::PGM_Raster (istream& in, boolean ascii) {
         in.gets(&buffer);
 #else
 	in.get(buffer, BUFSIZ);
+	in.get(newline);
 #endif
     } while (buffer[0] == '#');
 
@@ -1908,13 +1950,14 @@ OverlayRaster* OvImportCmd::PGM_Raster (istream& in, boolean ascii) {
           in.gets(&buffer);
 #else
 	  in.get(buffer, BUFSIZ);
+	  in.get(newline);
 #endif
           sscanf(buffer, "%d", &nrows);
     }
 #if __GNUG__<3 
     in.gets(&buffer);
 #else
-    in.get(buffer, BUFSIZ);
+    in.get(buffer, BUFSIZ, '\n');
 #endif
     int maxval;
     sscanf(buffer, "%d", &maxval);
@@ -2144,7 +2187,7 @@ FILE* OvImportCmd::Portable_Raster_Open(
     int& ncols, int& nrows, boolean& compressed, boolean& tiled, int& twidth, 
     int& theight
 ) {
-    FILE* file = fopen(pathname, "r");
+    FILE* file = fopen(pathname, "r+");
     file = CheckCompression(file, pathname, compressed);
   
     tiled = false;
@@ -2327,6 +2370,7 @@ OverlayRaster* OvImportCmd::PPM_Raster (istream& in, boolean ascii) {
 #else
     char buffer[BUFSIZ];
     in.get(buffer,BUFSIZ);
+    in.get(newline);
 #endif
 
     do { // CREATOR and other comments
@@ -2334,6 +2378,7 @@ OverlayRaster* OvImportCmd::PPM_Raster (istream& in, boolean ascii) {
         in.gets(&buffer);
 #else
 	in.get(buffer,BUFSIZ);
+	in.get(newline);
 #endif
     } while (buffer[0] == '#');
     int nrows, ncols;
@@ -2342,6 +2387,7 @@ OverlayRaster* OvImportCmd::PPM_Raster (istream& in, boolean ascii) {
     in.gets(&buffer);
 #else
     in.get(buffer,BUFSIZ);
+    in.get(newline);
 #endif
     int maxval;
     sscanf(buffer, "%d", &maxval);
@@ -2391,18 +2437,18 @@ GraphicComp* OvImportCmd::PNM_Image_Filter (
     ifstream in2;
     in2.rdbuf()->attach(outfd);
 #else
-    FILE* outfptr = fdopen(outfd, "w");
-    filebuf fbuf(outfptr, ios_base::out);
+    FILE* infptr = fdopen(outfd, "r");
+    filebuf fbuf(infptr, ios_base::in);
     istream in2(&fbuf);
 #endif
 
     comp = PNM_Image(in2);
 
-#if __GNUG__>=3
-    fclose(outfptr);
-#endif
     if(close(outfd)==-1)
       cerr << "error in parent closing last end of the pipes\n";
+#if __GNUG__>=3
+    if (infptr) fclose(infptr);
+#endif
   }
 
   return comp;
@@ -2467,11 +2513,6 @@ int OvImportCmd::Pipe_Filter (istream& in, const char* filter)
 #if __GNUG__<3
     ofstream out;
     out.rdbuf()->attach(pipe1[1]);
-#else
-    FILE* ofptr = fdopen(pipe1[1], "w");
-    filebuf fbuf(ofptr, ios_base::out);
-    ostream out(&fbuf);
-#endif
     char buffer[BUFSIZ];
     while (!in.eof() && in.good()) {
       in.read(buffer, BUFSIZ);
@@ -2479,8 +2520,13 @@ int OvImportCmd::Pipe_Filter (istream& in, const char* filter)
 	out.write(buffer, in.gcount());
     }
     out.flush();
-#if __GNUG__>=3
-    fclose(ofptr);
+#else
+    char buffer[BUFSIZ];
+    while (!in.eof() && in.good()) {
+      in.read(buffer, BUFSIZ);
+      if (!in.eof() || in.gcount())
+	write(pipe1[1], buffer, in.gcount());
+    }
 #endif
     if(close(pipe1[1])==-1)
       cerr << "error in child closing its output pipe\n";
@@ -2564,7 +2610,7 @@ GraphicComp* OvImportCmd::PBM_Image (const char* pathname) {
 
 Bitmap* OvImportCmd::PBM_Bitmap (const char* pathname) {
     Bitmap* bitmap = nil;
-    FILE* file = fopen(pathname, "r");
+    FILE* file = fopen(pathname, "r+");
     boolean compressed;
     file = CheckCompression(file, pathname, compressed);
 
@@ -2651,6 +2697,7 @@ Bitmap* OvImportCmd::PBM_Bitmap (istream& in) {
 #else
     char buffer[BUFSIZ];
     in.get(buffer,BUFSIZ);
+    in.get(newline);
 #endif
 
     boolean asciiflag = strncmp("P1", buffer, 2) == 0;
@@ -2660,6 +2707,7 @@ Bitmap* OvImportCmd::PBM_Bitmap (istream& in) {
         in.gets(&buffer);
 #else
 	in.get(buffer,BUFSIZ);
+	in.get(newline);
 #endif
     } while (buffer[0] == '#');
 
@@ -2669,6 +2717,7 @@ Bitmap* OvImportCmd::PBM_Bitmap (istream& in) {
           in.gets(&buffer);
 #else
 	  in.get(buffer,BUFSIZ);
+	  in.get(newline);
 #endif
           sscanf(buffer, "%d", &nrows);
     }
